Change Model#__delete__() to raise error when non-inserted object.

+++ ruby/lib/kwery/model.rb
@@ -166,6 +166,7 @@ module Kwery
     end
 
     def __delete__(query)
+      raise "Not inserted object." if @__old__.nil?
       query.delete(self.class.__table__, self.id)
       __deleted__(query)
     end
